+++ driver-core/src/main/java/com/datastax/driver/core/RequestHandler.java
@@ -513,10 +513,12 @@ class RequestHandler implements Connection.ResponseCallback {
     public void onTimeout(Connection connection, long latency) {
         Host queriedHost = current;
         try {
-            if (connection instanceof PooledConnection)
-                ((PooledConnection)connection).release();
+            // If a query times out, we consider that the host is unstable, so we defunct
+            // the connection to mark it down.
+            DriverException timeoutException = new DriverException("Timed out waiting for server response");
+            connection.defunct(timeoutException);
 
-            logError(connection.address, new DriverException("Timeout during read"));
+            logError(connection.address, timeoutException);
             retry(false, null);
         } catch (Exception e) {
             // This shouldn't happen, but if it does, we want to signal the callback, not let him hang indefinitively